channel-main: Fix monitors_align to not shuffle monitor order
Before this patch monitor_align was calling qsort directly on the
VDAgentMonConfig monitors array, but VDAgentMonConfig does not contain
an id, so the order matters!

This fixes (for example) 2 issues with having 3 windows/monitors on a row
numbered 1-3, ordered left-to-right as 1-2-3, and then changing the
ordering to 1-3-2:
1) Window 3 would be resized to the size of window 2, and window 2 would
   get resized to the size of window 3.
2) Dragging a window on monitor 1 over its right edge, makes the part over
   the edge show up on the right monitor, rather then on the middle.
This is happening because the agent is configuring qxl-1 (which is monitor 2)
with the monitors[1] data, which after the qsort contains the size and
coordinates of monitor 3.

Note this only happens with virt-viewer fixed to properly send window
coordinates, as before that all monitors had x and y set to 0 making the
sort a nop.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>